--- athena/optimize/__init__.py
@@ -1,4 +1,0 @@
-from athena.optimize.strategies.strategy import Strategy
-from athena.optimize.orders import Position, Trade, Portfolio
-
-__all__ = ["Strategy", "Position", "Trade", "Portfolio"]
