+++ athena/tradingtools/backtesting.py
@@ -1,4 +1,4 @@
-from athena.optimize import Strategy, Trade, Portfolio, Position
+from athena.tradingtools import Strategy, Trade, Portfolio, Position
 from athena.core.types import Coin, Signal
 from athena.core.interfaces import Fluctuations, Candle
 import datetime
