+++ athena/tradingtools/strategies/strategy.py
